refactor: move geniusyield parser to dex package; real deployment identifiers
