+++ dex/geniusyield/models.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	"github.com/blinklabs-io/gouroboros/cbor"
-	"github.com/blinklabs-io/shai/internal/common"
+	"github.com/blinklabs-io/shai/common"
 )
 
 // Protocol constants
