+++ dex/geniusyield/parser.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/blinklabs-io/gouroboros/cbor"
-	"github.com/blinklabs-io/shai/internal/common"
+	"github.com/blinklabs-io/shai/common"
 )
 
 // OrderState represents the parsed state of a Genius Yield order
